+++ src/plugins/discover/public/application/components/top_nav/open_options_popover.tsx
@@ -6,55 +6,15 @@
  * Side Public License, v 1.
  */
 
-import React from 'react';
+import React, { Suspense, lazy } from 'react';
 import ReactDOM from 'react-dom';
 import { I18nStart } from 'kibana/public';
-import { i18n } from '@kbn/i18n';
-import { FormattedMessage } from '@kbn/i18n/react';
-import { EuiSpacer, EuiLink, EuiText, EuiWrappingPopover } from '@elastic/eui';
-import { getServices } from '../../../kibana_services';
+import { EuiLoadingSpinner } from '@elastic/eui';
 import './open_options_popover.scss';
 
 let isOpen = false;
 
-interface OptionsPopoverProps {
-  onClose: () => void;
-  anchorElement: HTMLElement;
-}
-
-export function OptionsPopover(props: OptionsPopoverProps) {
-  const {
-    core: { uiSettings },
-    addBasePath,
-  } = getServices();
-  const isLegacy = uiSettings.get('doc_table:legacy');
-
-  const linkText = isLegacy
-    ? i18n.translate('discover.openOptionsPopover.switchToDataGridText', {
-        defaultMessage: 'Switch to data grid',
-      })
-    : i18n.translate('discover.openOptionsPopover.switchToLegacyText', {
-        defaultMessage: 'Switch to legacy table',
-      });
-
-  return (
-    <EuiWrappingPopover ownFocus button={props.anchorElement} closePopover={props.onClose} isOpen>
-      <div className="dscOptionsPopover">
-        <EuiLink href={addBasePath('/app/management/kibana/settings?query=Use legacy table')}>
-          {linkText}
-        </EuiLink>
-        <EuiSpacer size="s" />
-        <EuiText color="subdued" size="s">
-          <FormattedMessage
-            id="discover.topNav.openOptionsPopover.description"
-            defaultMessage="The new data grid layout includes better data sorting, drag-and-drop columns, and a full
-                screen view. Enable this option if you prefer to fall back to the legacy table."
-          />
-        </EuiText>
-      </div>
-    </EuiWrappingPopover>
-  );
-}
+const OptionsPopoverLazy = lazy(() => import('./options_popover'));
 
 export function openOptionsPopover({
   I18nContext,
@@ -79,7 +39,9 @@ export function openOptionsPopover({
 
   const element = (
     <I18nContext>
-      <OptionsPopover onClose={onClose} anchorElement={anchorElement} />
+      <Suspense fallback={<EuiLoadingSpinner />}>
+        <OptionsPopoverLazy onClose={onClose} anchorElement={anchorElement} />
+      </Suspense>
     </I18nContext>
   );
   ReactDOM.render(element, container);
